+++ Example/ProductsExample.php
@@ -113,11 +113,11 @@ try {
 	print('Product update failed. To confirm price difference use force token: ');
 	// get force token
 	$forceToken = $ex->getForceToken();
-	var_export($forceToken);
+	print($forceToken);
 	print(PHP_EOL);
 	// set token to the client args and repeat product update
-	$mpapiClient->setArgument(Product::ARG_FORCE_TOKEN, $forceToken);
-	$response = $products->put('pTU00_test', $product);
+	$updateWithForceToken = $products->put('pTU00_test', $product, null, $forceToken);
+	var_dump($updateWithForceToken);
 }
 
 // ####################################